+++ lib/src/model/ble_device.dart
@@ -1,5 +1,0 @@
-class BleDevice {
-  String name;
-  String id;
-  BleDevice({required this.name, required this.id});
-}