+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2
+
+- Added All Bluetooth Apis
+- Added Filter Options in request Device
+- fix read / write Characteristics
+- updated bluetooth plugin -> flutter_reactive_ble
+- updated scanDevice Dialog box UI for android and Ios both
+- minor fixes
+
 ## 0.0.1
 
 - initial release.